Move getPublic() to cql/cfg.go, remove 'cql generate nonce' type.
